+++ src/internalPages/ExtensionMain/MainScreen.tsx
@@ -21,10 +21,10 @@ interface Props extends WithStyles<typeof styles>{
 }
 
 const MainScreen:React.FC<Props> = ({classes}) => {
-  const [gatewayBaseURL, setGatewayBaseURL] = useState(ExtensionURIMap[ExtensionOptions.CloudlareCDN]);
+  const [gatewayBaseURL, setGatewayBaseURL] = useState(ExtensionURIMap[ExtensionOptions.CloudflareCDN]);
   const [okGatewayBaseURL, setOkGatewayBaseURL] = useState(false);
   const [showTexField, setShowTextField] = useState(false);
-  const [gatewayOption, setGateWayOption] = useState<ExtensionOptions>(ExtensionOptions.CloudlareCDN);
+  const [gatewayOption, setGateWayOption] = useState<ExtensionOptions>(ExtensionOptions.CloudflareCDN);
 
   useEffect(() => {
     chromeStorageSyncGet(StorageSyncKey.GatewayOption).then(option => {
@@ -94,7 +94,7 @@ const MainScreen:React.FC<Props> = ({classes}) => {
               value={gatewayOption}
               onChange={handleChange}
             >
-              <MenuItem value={ExtensionOptions.CloudlareCDN}>Cloudlare CDN</MenuItem>
+              <MenuItem value={ExtensionOptions.CloudflareCDN}>Cloudflare CDN</MenuItem>
               <MenuItem value={ExtensionOptions.InfuraAPI}>Infura API</MenuItem>
               <MenuItem value={ExtensionOptions.IPFSNetwork}>Directly from IPFS network</MenuItem>
               <MenuItem value={ExtensionOptions.Pinata}>Pinata IPFS network</MenuItem>