+++ src/types.ts
@@ -1,6 +1,6 @@
 
 export enum ExtensionOptions {
-  CloudlareCDN="Cloudlare CDN",
+  CloudflareCDN="Cloudflare CDN",
   InfuraAPI = "Infura API",
   IPFSNetwork = "Directly from IPFS network",
   Pinata = "Pinata",
@@ -16,7 +16,7 @@ export interface ExtensionURIMap {
 };
 
 export const ExtensionLabel: ExtensionOptionMessage = {
-  [ExtensionOptions.CloudlareCDN] : "Non-paranoid + fast response times",
+  [ExtensionOptions.CloudflareCDN] : "Non-paranoid + fast response times",
   [ExtensionOptions.InfuraAPI] : "Non-paranoid + fast response times",
   [ExtensionOptions.IPFSNetwork]: "Paranoid + slow response times",
   [ExtensionOptions.Pinata]: "Non-paranoid + fast response times",
@@ -24,7 +24,7 @@ export const ExtensionLabel: ExtensionOptionMessage = {
 };
 
 export const ExtensionURIMap: ExtensionURIMap = {
-  [ExtensionOptions.CloudlareCDN]: 'https://cloudflare-ipfs.com/',
+  [ExtensionOptions.CloudflareCDN]: 'https://cloudflare-ipfs.com/',
   [ExtensionOptions.InfuraAPI]: 'https://ipfs.infura.io/',
   [ExtensionOptions.IPFSNetwork]: 'https://gateway.ipfs.io/',
   [ExtensionOptions.Pinata]: 'https://abbfe6z95qov3d40hf6j30g7auo7afhp.mypinata.cloud/'